vcm: refactor vcm.calc.thermo and improve type hints (#1576)
* Refactor vertical coordinate funcs to new module
* Move hydrostatic_dz to vertical_coordinates.py
* Fix fv3fit thermo import
* Rename module
* Fix docs references
* Fix import
