+++ external/vcm/vcm/calc/thermo.py
@@ -1,188 +1,27 @@
 import numpy as np
 import xarray as xr
-from ..cubedsphere.constants import COORD_Z_CENTER, COORD_Z_OUTER
-
-
-# following are defined as in FV3GFS model (see FV3/fms/constants/constants.f90)
-_GRAVITY = 9.80665  # m /s2
-_RDGAS = 287.05  # J / K / kg
-_RVGAS = 461.5  # J / K / kg
-_LATENT_HEAT_VAPORIZATION_0_C = 2.5e6  # J/kg
-_LATENT_HEAT_FUSION = 3.3358e5  # J/kg
-_SPECIFIC_ENTHALPY_LIQUID = 4185.5
-_SPECIFIC_ENTHALPY_VAP0R = 1846
-_SPECIFIC_HEAT_CONST_PRESSURE = 1004
-_FREEZING_TEMPERATURE = 273.15
-_POISSON_CONST = 0.2854
-_EARTH_RADIUS = 6.3712e6  # m
-
-_DEFAULT_SURFACE_TEMPERATURE = _FREEZING_TEMPERATURE + 15
-
-_TOA_PRESSURE = 300.0  # Pa
-_REFERENCE_SURFACE_PRESSURE = 100000  # reference pressure for potential temp [Pa]
-_REVERSE = slice(None, None, -1)
-
-_SEC_PER_DAY = 86400
-_KG_M2S_TO_MM_DAY = (1e3 * 86400) / 997.0
-_KG_M2_TO_MM = 1000.0 / 997
-
-
-def mass_integrate(da, delp, dim=COORD_Z_CENTER):
-    return (da * delp / _GRAVITY).sum(dim)
+from .vertical_coordinate import mass_integrate
+from .constants import (
+    _GRAVITY,
+    _REFERENCE_SURFACE_PRESSURE,
+    _POISSON_CONST,
+    _RDGAS,
+    _LATENT_HEAT_FUSION,
+    _LATENT_HEAT_VAPORIZATION_0_C,
+    _SPECIFIC_ENTHALPY_LIQUID,
+    _SPECIFIC_ENTHALPY_VAP0R,
+    _FREEZING_TEMPERATURE,
+    _DEFAULT_SURFACE_TEMPERATURE,
+    _KG_M2_TO_MM,
+    _KG_M2S_TO_MM_DAY,
+    _SEC_PER_DAY,
+    _SPECIFIC_HEAT_CONST_PRESSURE,
+    _EARTH_RADIUS,
+)
 
 
 def potential_temperature(P, T):
     return T * (_REFERENCE_SURFACE_PRESSURE / P) ** _POISSON_CONST
-
-
-def pressure_at_interface(
-    delp, toa_pressure=_TOA_PRESSURE, dim_center=COORD_Z_CENTER, dim_outer=COORD_Z_OUTER
-):
-    """ Compute pressure at layer interfaces
-
-    Args:
-        delp (xr.DataArray): pressure thicknesses
-        toa_pressure (float, optional): pressure at the top of atmosphere.
-            Defaults to 300Pa.
-        dim_center (str, optional): name of vertical dimension for delp.
-            Defaults to "pfull".
-        dim_outer (str, optional): name of vertical dimension for output pressure.
-            Defaults to "phalf".
-
-    Returns:
-        xr.DataArray: atmospheric pressure at layer interfaces
-    """
-    top = xr.full_like(delp.isel({dim_center: [0]}), toa_pressure).variable
-    delp_with_top = top.concat([top, delp.variable], dim=dim_center)
-    pressure = delp_with_top.cumsum(dim_center)
-    return _add_coords_to_interface_variable(
-        pressure, delp, dim_center=dim_center
-    ).rename({dim_center: dim_outer})
-
-
-def height_at_interface(dz, phis, dim_center=COORD_Z_CENTER, dim_outer=COORD_Z_OUTER):
-    """ Compute geopotential height at layer interfaces
-
-    Args:
-        dz (xr.DataArray): layer height thicknesses
-        phis (xr.DataArray): surface geopotential
-        dim_center (str, optional): name of vertical dimension for dz.
-            Defaults to "pfull".
-        dim_outer (str, optional): name of vertical dimension for output heights.
-            Defaults to "phalf".
-
-    Returns:
-        xr.DataArray: height at layer interfaces
-    """
-    bottom = phis.broadcast_like(dz.isel({dim_center: [0]})) / _GRAVITY
-    dzv = -dz.variable  # dz is negative in model
-    dz_with_bottom = dzv.concat([dzv, bottom], dim=dim_center)
-    height = (
-        dz_with_bottom.isel({dim_center: _REVERSE})
-        .cumsum(dim_center)
-        .isel({dim_center: _REVERSE})
-    )
-    return _add_coords_to_interface_variable(height, dz, dim_center=dim_center).rename(
-        {dim_center: dim_outer}
-    )
-
-
-def _add_coords_to_interface_variable(
-    dv_outer: xr.Variable, da_center: xr.DataArray, dim_center: str = COORD_Z_CENTER
-):
-    """Assign all coords except vertical from da_center to dv_outer """
-    if dim_center in da_center.coords:
-        return xr.DataArray(dv_outer, coords=da_center.drop_vars(dim_center).coords)
-    else:
-        return xr.DataArray(dv_outer, coords=da_center.coords)
-
-
-def pressure_at_midpoint(delp, toa_pressure=_TOA_PRESSURE, dim=COORD_Z_CENTER):
-    """ Compute pressure at layer midpoints by linear interpolation
-
-    Args:
-        delp (xr.DataArray): pressure thicknesses
-        toa_pressure (float, optional): pressure at the top of atmosphere.
-            Defaults to 300Pa.
-        dim (str, optional): name of vertical dimension for delp. Defaults to "pfull".
-
-    Returns:
-        xr.DataArray: atmospheric pressure at layer midpoints
-    """
-    pi = pressure_at_interface(delp, toa_pressure=toa_pressure, dim_center=dim)
-    return _interface_to_midpoint(pi, dim_center=dim)
-
-
-def height_at_midpoint(dz, phis, dim=COORD_Z_CENTER):
-    """ Compute geopotential height at layer midpoints by linear interpolation
-
-    Args:
-        dz (xr.DataArray): layer height thicknesses
-        phis (xr.DataArray): surface geopotential
-        dim (str, optional): name of vertical dimension for dz. Defaults to "pfull".
-
-    Returns:
-        xr.DataArray: height at layer midpoints
-    """
-    zi = height_at_interface(dz, phis, dim_center=dim)
-    return _interface_to_midpoint(zi, dim_center=dim)
-
-
-def _interface_to_midpoint(da, dim_center=COORD_Z_CENTER, dim_outer=COORD_Z_OUTER):
-    da_mid = (
-        da.isel({dim_outer: slice(0, -1)}) + da.isel({dim_outer: slice(1, None)})
-    ) / 2
-    return da_mid.rename({dim_outer: dim_center})
-
-
-def pressure_at_midpoint_log(delp, toa_pressure=_TOA_PRESSURE, dim=COORD_Z_CENTER):
-    """ Compute pressure at layer midpoints following Eq. 3.17 of Simmons
-    and Burridge (1981), MWR.
-
-    Args:
-        delp (xr.DataArray): pressure thicknesses
-        toa_pressure (float, optional): pressure at the top of atmosphere.
-            Defaults to 300Pa.
-        dim (str, optional): name of vertical dimension for delp. Defaults to "pfull".
-
-    Returns:
-        xr.DataArray: atmospheric pressure at layer midpoints
-    """
-    pi = pressure_at_interface(
-        delp, toa_pressure=toa_pressure, dim_center=dim, dim_outer=dim
-    )
-    dlogp = np.log(pi).diff(dim)
-    output = delp / dlogp
-    # ensure that output chunks are the same
-    # pressure at interface adds a single chunk at the model surface
-    if delp.chunks:
-        original_chunk_size = delp.chunks[delp.get_axis_num(dim)]
-        output = output.chunk({dim: original_chunk_size})
-
-    return output
-
-
-def hydrostatic_dz(T, q, delp, dim=COORD_Z_CENTER):
-    """ Compute layer thickness assuming hydrostatic balance
-
-    Args:
-        T (xr.DataArray): temperature
-        q (xr.DataArray): specific humidity
-        delp (xr.DataArray): pressure thickness
-        dim (str, optional): name of vertical dimension. Defaults to "pfull".
-
-    Returns:
-        xr.DataArray: layer thicknesses dz
-    """
-    pi = pressure_at_interface(delp, dim_center=dim, dim_outer=dim)
-    tv = T * (1 + (_RVGAS / _RDGAS - 1) * q)
-    dlogp = np.log(pi).diff(dim)
-    return -dlogp * _RDGAS * tv / _GRAVITY
-
-
-def dz_and_top_to_phis(top_height, dz, dim=COORD_Z_CENTER):
-    """ Compute surface geopotential from model top height and layer thicknesses"""
-    return _GRAVITY * (top_height + dz.sum(dim=dim))
 
 
 def latent_heat_vaporization(T):
@@ -192,17 +31,17 @@ def latent_heat_vaporization(T):
 
 
 def net_heating(
-    dlw_sfc,
-    dsw_sfc,
-    ulw_sfc,
-    ulw_toa,
-    usw_sfc,
-    usw_toa,
-    dsw_toa,
-    shf,
-    surface_rain_rate,
-    surface_temperature=_FREEZING_TEMPERATURE + 10,
-):
+    dlw_sfc: xr.DataArray,
+    dsw_sfc: xr.DataArray,
+    ulw_sfc: xr.DataArray,
+    ulw_toa: xr.DataArray,
+    usw_sfc: xr.DataArray,
+    usw_toa: xr.DataArray,
+    dsw_toa: xr.DataArray,
+    shf: xr.DataArray,
+    surface_rain_rate: xr.DataArray,
+    surface_temperature: float = _FREEZING_TEMPERATURE + 10,
+) -> xr.DataArray:
     """A dataarray implementation of ``net_heating_from_dataset``
 
     All argument except for ``cond_int`` should be in W/m2. cond_int is in units
@@ -230,8 +69,8 @@ def net_heating(
 
 
 def latent_heat_flux_to_evaporation(
-    lhf, surface_temperature=_DEFAULT_SURFACE_TEMPERATURE
-):
+    lhf: xr.DataArray, surface_temperature: float = _DEFAULT_SURFACE_TEMPERATURE
+) -> xr.DataArray:
     """Compute evaporation from latent heat flux
 
     Args:
@@ -267,32 +106,10 @@ def surface_evaporation_mm_day_from_latent_heat_flux(
     return surface_evaporation
 
 
-def net_precipitation(lhf, prate):
+def net_precipitation(lhf: xr.DataArray, prate: xr.DataArray) -> xr.DataArray:
     da = (prate - latent_heat_flux_to_evaporation(lhf)) * _SEC_PER_DAY
     da.attrs = {"long_name": "net precipitation from model physics", "units": "mm/day"}
     return da
-
-
-def surface_pressure_from_delp(
-    delp: xr.DataArray, p_toa: float = 300.0, vertical_dim: str = "z"
-) -> xr.DataArray:
-    """Compute surface pressure from delp in Pa
-    
-    Args:
-        delp: DataArray of pressure layer thicknesses in Pa
-        p_toa: Top of atmosphere pressure in Pa; optional, defaults to 300
-        vertical_dim: Name of vertical dimension; defaults to 'z'
-        
-    Returns:
-        surface_pressure: DataArray of surface pressure in Pa
-    """
-
-    surface_pressure = delp.sum(dim=vertical_dim) + p_toa
-    surface_pressure = surface_pressure.assign_attrs(
-        {"long_name": "surface pressure", "units": "Pa"}
-    )
-
-    return surface_pressure
 
 
 def total_water(
